+++ test/Columns.test.js
@@ -32,6 +32,10 @@ const testColumnsMethod = (testConfig) => {
      * }
      */
     const reducedTests = testConfig.tests.reduce((acc, test) => {
+      if (true === test.skip) {
+        return acc;
+      }
+
       // TODO: Use Object.values(test) once Node v6 support is dropped.
       const testValues = Object.keys(test).map(key => test[key]);
       return [...acc, testValues];
@@ -39,12 +43,76 @@ const testColumnsMethod = (testConfig) => {
 
     test.each(reducedTests)(
       '%s',
-      (theTest) => {
-        expect(theTest.actual).toEqual(theTest.expected);
+      (theTest, actual, expected) => {
+        expect(actual).toEqual(expected);
       },
     );
   });
 };
+
+describe('Checking values passed to Columns.buildCalcFunction()', () => {
+  describe('All options', () => {
+    const instance = new Columns(allValues);
+    jest.spyOn(instance, 'buildCalcFunction');
+
+    test('Single column', () => {
+      instance.spanCalc(1);
+      expect(instance.buildCalcFunction).toHaveBeenCalledWith('100vw', 1, 0);
+      expect(instance.buildCalcFunction).toHaveBeenCalledWith('75rem', 1, 0);
+    });
+
+    test('Fractional columns (greater than 2)', () => {
+      instance.spanCalc(2.5);
+      expect(instance.buildCalcFunction).toHaveBeenCalledWith('100vw', 2.5, 2);
+      expect(instance.buildCalcFunction).toHaveBeenCalledWith('75rem', 2.5, 2);
+    });
+  });
+
+  test('Omits a `full` value with no `siteMax` option', () => {
+    const instance = new Columns(edgeGap);
+    jest.spyOn(instance, 'buildCalcFunction');
+
+    instance.spanCalc(2.5);
+    expect(instance.buildCalcFunction).toHaveBeenCalledWith('100vw', 2.5, 2);
+  });
+});
+
+describe('Get calc functions from buildCalcFunction()', () => {
+  describe('All options', () => {
+    const instance = new Columns(allValues); // eslint-disable-line
+
+    test('Single column', () => {
+      const expected = {
+        fluid: 'calc(6.25vw - 4px - 0.5859rem)',
+        full: 'calc(4.1016rem - 4px)',
+      };
+
+      expect(instance.buildCalcFunction('100vw', 1, 0)).toEqual(expected.fluid);
+      expect(instance.buildCalcFunction('75rem', 1, 0)).toEqual(expected.full);
+    });
+
+    test('Fractional columns (greater than 2)', () => {
+      const expected = {
+        fluid: 'calc(15.625vw - 10px - 0.2148rem)',
+        full: 'calc(11.504rem - 10px)',
+      };
+
+      expect(instance.buildCalcFunction('100vw', 2.5, 2)).toEqual(expected.fluid);
+      expect(instance.buildCalcFunction('75rem', 2.5, 2)).toEqual(expected.full);
+    });
+  });
+
+  test('Omits a `full` value with no `siteMax` option', () => {
+    const instance = new Columns(edgeGap);
+
+    const expected = {
+      fluid: 'calc(8.3333vw - 0.1667rem - 9.1667px)',
+    };
+
+    expect(instance.buildCalcFunction('100vw', 1, 0)).toEqual(expected.fluid);
+    expect(undefined).toEqual(expected.full);
+  });
+});
 
 /**
  * Calculate the shared gap amount to be removed from each column.
@@ -86,56 +154,70 @@ testColumnsMethod({
       description: 'All options: single column',
       actual: new Columns(allValues).spanCalc(1),
       expected: {
-        fluid: 'calc((100vw - 32px * 2) / 16 - 0.5859rem)',
-        full: 'calc((75rem - 32px * 2) / 16 - 0.5859rem)',
+        // calc((100vw - 32px * 2) / 16 - 0.5859rem)
+        fluid: 'calc(6.25vw - 4px - 0.5859rem)',
+        // calc((75rem - 32px * 2) / 16 - 0.5859rem)
+        full: 'calc(4.1016rem - 4px)',
       },
     },
     {
       description: 'All options: two columns',
       actual: new Columns(allValues).spanCalc(2),
       expected: {
-        fluid: 'calc((((100vw - 32px * 2) / 16 - 0.5859rem) * 2) + 0.625rem)',
-        full: 'calc((((75rem - 32px * 2) / 16 - 0.5859rem) * 2) + 0.625rem)',
+        // calc((((100vw - 32px * 2) / 16 - 0.5859rem) * 2) + 0.625rem)
+        fluid: 'calc(12.5vw - 8px - 0.5468rem)',
+        // calc((((75rem - 32px * 2) / 16 - 0.5859rem) * 2) + 0.625rem)
+        full: 'calc(8.8282rem - 8px)',
       },
     },
     {
       description: 'All options: three columns',
       actual: new Columns(allValues).spanCalc(4),
       expected: {
-        fluid: 'calc((((100vw - 32px * 2) / 16 - 0.5859rem) * 4) + 0.625rem * 3)',
-        full: 'calc((((75rem - 32px * 2) / 16 - 0.5859rem) * 4) + 0.625rem * 3)',
+        // calc((((100vw - 32px * 2) / 16 - 0.5859rem) * 4) + 0.625rem * 3)
+        fluid: 'calc(25vw - 16px - 0.4686rem)',
+        // calc((((75rem - 32px * 2) / 16 - 0.5859rem) * 4) + 0.625rem * 3)
+        full: 'calc(18.2814rem - 16px)',
       },
     },
     {
       description: 'All options: negative columns',
       actual: new Columns(allValues).spanCalc(-4),
       expected: {
-        fluid: 'calc((((100vw - 32px * 2) / 16 - 0.5859rem) * -4) + 0.625rem * -3)',
-        full: 'calc((((75rem - 32px * 2) / 16 - 0.5859rem) * -4) + 0.625rem * -3)',
+        // calc((((100vw - 32px * 2) / 16 - 0.5859rem) * -4) + 0.625rem * -3)
+        fluid: 'calc(-25vw + 16px + 0.4686rem)',
+        // calc((((75rem - 32px * 2) / 16 - 0.5859rem) * -4) + 0.625rem * -3)
+        full: 'calc(-18.2814rem + 16px)',
       },
     },
     {
       description: 'All options: fractional columns (less than 1)',
       actual: new Columns(allValues).spanCalc(0.5),
       expected: {
-        fluid: 'calc(((100vw - 32px * 2) / 16 - 0.5859rem) * 0.5)',
-        full: 'calc(((75rem - 32px * 2) / 16 - 0.5859rem) * 0.5)',
+        // calc(((100vw - 32px * 2) / 16 - 0.5859rem) * 0.5)
+        fluid: 'calc(3.125vw - 2px - 0.293rem)',
+        // calc(((75rem - 32px * 2) / 16 - 0.5859rem) * 0.5)
+        full: 'calc(2.0508rem - 2px)',
       },
     },
     {
       description: 'All options: fractional columns (greater than 1)',
       actual: new Columns(allValues).spanCalc(1.75),
       expected: {
-        fluid: 'calc((((100vw - 32px * 2) / 16 - 0.5859rem) * 1.75) + 0.625rem)',
-        full: 'calc((((75rem - 32px * 2) / 16 - 0.5859rem) * 1.75) + 0.625rem)',
+        // calc((((100vw - 32px * 2) / 16 - 0.5859rem) * 1.75) + 0.625rem)
+        fluid: 'calc(10.9375vw - 7px - 0.4003rem)',
+        // calc((((75rem - 32px * 2) / 16 - 0.5859rem) * 1.75) + 0.625rem)
+        full: 'calc(7.8028rem - 7px)',
       },
     },
     {
       description: 'All options: fractional columns (greater than 2)',
       actual: new Columns(allValues).spanCalc(2.5),
       expected: {
-        fluid: 'calc((((100vw - 32px * 2) / 16 - 0.5859rem) * 2.5) + 0.625rem * 2)',
-        full: 'calc((((75rem - 32px * 2) / 16 - 0.5859rem) * 2.5) + 0.625rem * 2)',
+        // calc((((100vw - 32px * 2) / 16 - 0.5859rem) * 2.5) + 0.625rem * 2)
+        fluid: 'calc(15.625vw - 10px - 0.2148rem)',
+        // calc((((75rem - 32px * 2) / 16 - 0.5859rem) * 2.5) + 0.625rem * 2)
+        full: 'calc(11.504rem - 10px)',
       },
     },
     // ---------- No siteMax
@@ -143,7 +225,8 @@ testColumnsMethod({
       description: 'Omits a `full` value with no `siteMax` option',
       actual: new Columns(edgeGap).spanCalc(1),
       expected: {
-        fluid: 'calc((100vw - 1rem * 2) / 12 - 9.1667px)',
+        // calc((100vw - 1rem * 2) / 12 - 9.1667px)
+        fluid: 'calc(8.3333vw - 0.1667rem - 9.1667px)',
       },
     },
     // ---------- No gap
@@ -151,16 +234,20 @@ testColumnsMethod({
       description: 'Omits shared gap for single column with no `gap` option',
       actual: new Columns(edgeSiteMax).spanCalc(1),
       expected: {
-        fluid: 'calc((100vw - 1.25rem * 2) / 16)',
-        full: 'calc((1024px - 1.25rem * 2) / 16)',
+        // calc((100vw - 1.25rem * 2) / 16)
+        fluid: 'calc(6.25vw - 0.1563rem)',
+        // calc((1024px - 1.25rem * 2) / 16)
+        full: 'calc(64px - 0.1563rem)',
       },
     },
     {
       description: 'Omits the gap addition wtih no `gap` option',
       actual: new Columns(edgeSiteMax).spanCalc(2),
       expected: {
-        fluid: 'calc(((100vw - 1.25rem * 2) / 16) * 2)',
-        full: 'calc(((1024px - 1.25rem * 2) / 16) * 2)',
+        // calc(((100vw - 1.25rem * 2) / 16) * 2)
+        fluid: 'calc(12.5vw - 0.3125rem)',
+        // calc(((1024px - 1.25rem * 2) / 16) * 2)
+        full: 'calc(128px - 0.3125rem)',
       },
     },
     // ---------- No edge
@@ -168,8 +255,10 @@ testColumnsMethod({
       description: 'Omits the edge subtraction with no `edge` option',
       actual: new Columns(gapSiteMax).spanCalc(1),
       expected: {
-        fluid: 'calc(100vw / 16 - 14.0625px)',
-        full: 'calc(60rem / 16 - 14.0625px)',
+        // calc(100vw / 16 - 14.0625px)
+        fluid: 'calc(6.25vw - 14.0625px)',
+        // calc(60rem / 16 - 14.0625px)
+        full: 'calc(3.75rem - 14.0625px)',
       },
     },
     // ---------- Edges only
@@ -177,7 +266,8 @@ testColumnsMethod({
       description: 'Omits undeclared values from span ouput: `edge` only',
       actual: new Columns(edgeOnly).spanCalc(1),
       expected: {
-        fluid: 'calc((100vw - 20px * 2) / 12)',
+        // calc((100vw - 20px * 2) / 12)
+        fluid: 'calc(8.3333vw - 3.3333px)',
       },
     },
     // ---------- Gap only
@@ -185,7 +275,8 @@ testColumnsMethod({
       description: 'Omits undeclared values from span ouput: `gap` only',
       actual: new Columns(gapOnly).spanCalc(1),
       expected: {
-        fluid: 'calc(100vw / 12 - 0.8594rem)',
+        // calc(100vw / 12 - 0.8594rem)
+        fluid: 'calc(8.3333vw - 0.8594rem)',
       },
     },
     // ---------- siteMax only
@@ -193,16 +284,20 @@ testColumnsMethod({
       description: 'Omits undeclared values from span ouput: `siteMax` only',
       actual: new Columns(siteMaxOnly).spanCalc(1),
       expected: {
-        fluid: 'calc(100vw / 16)',
-        full: 'calc(1200px / 16)',
+        // calc(100vw / 16)
+        fluid: '6.25vw',
+        // calc(1200px / 16)
+        full: '75px',
       },
     },
     {
       description: 'Omits undeclared values from span ouput: `edge` only (multiple columns)',
       actual: new Columns(siteMaxOnly).spanCalc(5),
       expected: {
-        fluid: 'calc((100vw / 16) * 5)',
-        full: 'calc((1200px / 16) * 5)',
+        // calc((100vw / 16) * 5)
+        fluid: '31.25vw',
+        // calc((1200px / 16) * 5)
+        full: '375px',
       },
     },
     // ---------- Columns only
@@ -210,14 +305,16 @@ testColumnsMethod({
       description: 'Omits undeclared values from span ouput: `columns` only',
       actual: new Columns(columnsOnly).spanCalc(1),
       expected: {
-        fluid: 'calc(100vw / 12)',
+        // calc(100vw / 12)
+        fluid: '8.3333vw',
       },
     },
     {
       description: 'Omits undeclared values from span ouput: `columns` only (multiple columns)',
       actual: new Columns(columnsOnly).spanCalc(3),
       expected: {
-        fluid: 'calc((100vw / 12) * 3)',
+        // calc((100vw / 12) * 3)
+        fluid: '25vw',
       },
     },
     // ---------- Custom Properties
@@ -225,8 +322,10 @@ testColumnsMethod({
       description: 'Custom properties used in option values',
       actual: new Columns(customProperties).spanCalc(3),
       expected: {
-        fluid: 'calc((((100vw - var(--edge) * 2) / var(--columns) - (var(--gap) / var(--columns) * (var(--columns) - 1))) * 3) + var(--gap) * 2)',
-        full: 'calc((((90rem - var(--edge) * 2) / var(--columns) - (var(--gap) / var(--columns) * (var(--columns) - 1))) * 3) + var(--gap) * 2)',
+        // calc((((100vw - var(--edge) * 2) / var(--columns) - (var(--gap) / var(--columns) * (var(--columns) - 1))) * 3) + var(--gap) * 2)
+        fluid: 'calc((100vw - var(--edge)*2)/var(--columns)*3 - var(--gap)/var(--columns)*(var(--columns) - 1)*3 + var(--gap)*2)',
+        // calc((((90rem - var(--edge) * 2) / var(--columns) - (var(--gap) / var(--columns) * (var(--columns) - 1))) * 3) + var(--gap) * 2)
+        full: 'calc((90rem - var(--edge)*2)/var(--columns)*3 - var(--gap)/var(--columns)*(var(--columns) - 1)*3 + var(--gap)*2)',
       },
     },
   ],
@@ -243,56 +342,70 @@ testColumnsMethod({
       description: 'All options: single column',
       actual: new Columns(allValues).offsetCalc(1),
       expected: {
-        fluid: 'calc(((100vw - 32px * 2) / 16 - 0.5859rem) + 0.625rem)',
-        full: 'calc(((75rem - 32px * 2) / 16 - 0.5859rem) + 0.625rem)',
+        // calc(((100vw - 32px * 2) / 16 - 0.5859rem) + 0.625rem)
+        fluid: 'calc(6.25vw - 4px + 0.0391rem)',
+        // calc(((75rem - 32px * 2) / 16 - 0.5859rem) + 0.625rem)
+        full: 'calc(4.7266rem - 4px)',
       },
     },
     {
       description: 'All options: two columns',
       actual: new Columns(allValues).offsetCalc(2),
       expected: {
-        fluid: 'calc((((100vw - 32px * 2) / 16 - 0.5859rem) * 2) + 0.625rem * 2)',
-        full: 'calc((((75rem - 32px * 2) / 16 - 0.5859rem) * 2) + 0.625rem * 2)',
+        // calc((((100vw - 32px * 2) / 16 - 0.5859rem) * 2) + 0.625rem * 2)
+        fluid: 'calc(12.5vw - 8px + 0.0782rem)',
+        // calc((((75rem - 32px * 2) / 16 - 0.5859rem) * 2) + 0.625rem * 2)
+        full: 'calc(9.4532rem - 8px)',
       },
     },
     {
       description: 'All options: three columns',
       actual: new Columns(allValues).offsetCalc(3),
       expected: {
-        fluid: 'calc((((100vw - 32px * 2) / 16 - 0.5859rem) * 3) + 0.625rem * 3)',
-        full: 'calc((((75rem - 32px * 2) / 16 - 0.5859rem) * 3) + 0.625rem * 3)',
+        // calc((((100vw - 32px * 2) / 16 - 0.5859rem) * 3) + 0.625rem * 3)
+        fluid: 'calc(18.75vw - 12px + 0.1173rem)',
+        // calc((((75rem - 32px * 2) / 16 - 0.5859rem) * 3) + 0.625rem * 3)
+        full: 'calc(14.1798rem - 12px)',
       },
     },
     {
       description: 'All options: negative columns',
       actual: new Columns(allValues).offsetCalc(-4),
       expected: {
-        fluid: 'calc((((100vw - 32px * 2) / 16 - 0.5859rem) * -4) + 0.625rem * -4)',
-        full: 'calc((((75rem - 32px * 2) / 16 - 0.5859rem) * -4) + 0.625rem * -4)',
+        // calc((((100vw - 32px * 2) / 16 - 0.5859rem) * -4) + 0.625rem * -4)
+        fluid: 'calc(-25vw + 16px - 0.1564rem)',
+        // calc((((75rem - 32px * 2) / 16 - 0.5859rem) * -4) + 0.625rem * -4)
+        full: 'calc(-18.9064rem + 16px)',
       },
     },
     {
       description: 'All options: fractional columns (less than 1)',
       actual: new Columns(allValues).offsetCalc(0.75),
       expected: {
-        fluid: 'calc(((100vw - 32px * 2) / 16 - 0.5859rem) * 0.75)',
-        full: 'calc(((75rem - 32px * 2) / 16 - 0.5859rem) * 0.75)',
+        // calc(((100vw - 32px * 2) / 16 - 0.5859rem) * 0.75)
+        fluid: 'calc(4.6875vw - 3px - 0.4394rem)',
+        // calc(((75rem - 32px * 2) / 16 - 0.5859rem) * 0.75)
+        full: 'calc(3.0762rem - 3px)',
       },
     },
     {
       description: 'All options: fractional columns (greater than 1)',
       actual: new Columns(allValues).offsetCalc(1.5),
       expected: {
-        fluid: 'calc((((100vw - 32px * 2) / 16 - 0.5859rem) * 1.5) + 0.625rem)',
-        full: 'calc((((75rem - 32px * 2) / 16 - 0.5859rem) * 1.5) + 0.625rem)',
+        // calc((((100vw - 32px * 2) / 16 - 0.5859rem) * 1.5) + 0.625rem)
+        fluid: 'calc(9.375vw - 6px - 0.2538rem)',
+        // calc((((75rem - 32px * 2) / 16 - 0.5859rem) * 1.5) + 0.625rem)
+        full: 'calc(6.7774rem - 6px)',
       },
     },
     {
       description: 'All options: fractional columns (greater than 2)',
       actual: new Columns(allValues).offsetCalc(2.075),
       expected: {
-        fluid: 'calc((((100vw - 32px * 2) / 16 - 0.5859rem) * 2.075) + 0.625rem * 2)',
-        full: 'calc((((75rem - 32px * 2) / 16 - 0.5859rem) * 2.075) + 0.625rem * 2)',
+        // calc((((100vw - 32px * 2) / 16 - 0.5859rem) * 2.075) + 0.625rem * 2)
+        fluid: 'calc(12.9688vw - 8.3px + 0.0343rem)',
+        // calc((((75rem - 32px * 2) / 16 - 0.5859rem) * 2.075) + 0.625rem * 2)
+        full: 'calc(9.7608rem - 8.3px)',
       },
     },
     // ---------- No siteMax
@@ -300,7 +413,8 @@ testColumnsMethod({
       description: 'Omits a `full` value with no `siteMax` option',
       actual: new Columns(edgeGap).offsetCalc(1),
       expected: {
-        fluid: 'calc(((100vw - 1rem * 2) / 12 - 9.1667px) + 10px)',
+        // calc(((100vw - 1rem * 2) / 12 - 9.1667px) + 10px)
+        fluid: 'calc(8.3333vw - 0.1667rem + 0.8333px)',
       },
     },
     // ---------- No gap
@@ -308,16 +422,20 @@ testColumnsMethod({
       description: 'Omits shared gap for single column with no `gap` option',
       actual: new Columns(edgeSiteMax).offsetCalc(1),
       expected: {
-        fluid: 'calc((100vw - 1.25rem * 2) / 16)',
-        full: 'calc((1024px - 1.25rem * 2) / 16)',
+        // calc((100vw - 1.25rem * 2) / 16)
+        fluid: 'calc(6.25vw - 0.1563rem)',
+        // calc((1024px - 1.25rem * 2) / 16)
+        full: 'calc(64px - 0.1563rem)',
       },
     },
     {
       description: 'Omits the gap addition wtih no `gap` option',
       actual: new Columns(edgeSiteMax).offsetCalc(2),
       expected: {
-        fluid: 'calc(((100vw - 1.25rem * 2) / 16) * 2)',
-        full: 'calc(((1024px - 1.25rem * 2) / 16) * 2)',
+        // calc(((100vw - 1.25rem * 2) / 16) * 2)
+        fluid: 'calc(12.5vw - 0.3125rem)',
+        // calc(((1024px - 1.25rem * 2) / 16) * 2)
+        full: 'calc(128px - 0.3125rem)',
       },
     },
     // ---------- No edge
@@ -325,8 +443,10 @@ testColumnsMethod({
       description: 'Omits the edge subtraction with no `edge` option',
       actual: new Columns(gapSiteMax).offsetCalc(1),
       expected: {
-        fluid: 'calc((100vw / 16 - 14.0625px) + 15px)',
-        full: 'calc((60rem / 16 - 14.0625px) + 15px)',
+        // calc((100vw / 16 - 14.0625px) + 15px)
+        fluid: 'calc(6.25vw + 0.9375px)',
+        // calc((60rem / 16 - 14.0625px) + 15px)
+        full: 'calc(3.75rem + 0.9375px)',
       },
     },
     // ---------- Edges only
@@ -334,14 +454,16 @@ testColumnsMethod({
       description: 'Omits undeclared values from span ouput: `edge` only',
       actual: new Columns(edgeOnly).offsetCalc(1),
       expected: {
-        fluid: 'calc((100vw - 20px * 2) / 12)',
+        // calc((100vw - 20px * 2) / 12)
+        fluid: 'calc(8.3333vw - 3.3333px)',
       },
     },
     {
       description: 'Omits undeclared values from span ouput: `edge` only (multiple columns)',
       actual: new Columns(edgeOnly).offsetCalc(2),
       expected: {
-        fluid: 'calc(((100vw - 20px * 2) / 12) * 2)',
+        // calc(((100vw - 20px * 2) / 12) * 2)
+        fluid: 'calc(16.6667vw - 6.6667px)',
       },
     },
     // ---------- Gap only
@@ -349,7 +471,8 @@ testColumnsMethod({
       description: 'Omits undeclared values from span ouput: `gap` only',
       actual: new Columns(gapOnly).offsetCalc(1),
       expected: {
-        fluid: 'calc((100vw / 12 - 0.8594rem) + 0.9375rem)',
+        // calc((100vw / 12 - 0.8594rem) + 0.9375rem)
+        fluid: 'calc(8.3333vw + 0.0781rem)',
       },
     },
     // ---------- siteMax only
@@ -357,8 +480,10 @@ testColumnsMethod({
       description: 'Omits undeclared values from span ouput: `siteMax` only',
       actual: new Columns(siteMaxOnly).offsetCalc(1),
       expected: {
-        fluid: 'calc(100vw / 16)',
-        full: 'calc(1200px / 16)',
+        // calc(100vw / 16)
+        fluid: '6.25vw',
+        // calc(1200px / 16)
+        full: '75px',
       },
     },
     // ---------- Columns only
@@ -366,14 +491,16 @@ testColumnsMethod({
       description: 'Omits undeclared values from span ouput: `columns` only',
       actual: new Columns(columnsOnly).offsetCalc(1),
       expected: {
-        fluid: 'calc(100vw / 12)',
+        // calc(100vw / 12)
+        fluid: '8.3333vw',
       },
     },
     {
       description: 'Omits undeclared values from span ouput: `columns` only (multiple columns)',
       actual: new Columns(columnsOnly).offsetCalc(4),
       expected: {
-        fluid: 'calc((100vw / 12) * 4)',
+        // calc((100vw / 12) * 4)
+        fluid: '33.3333vw',
       },
     },
     // ---------- Custom Properties
@@ -381,8 +508,10 @@ testColumnsMethod({
       description: 'Custom properties used in option values',
       actual: new Columns(customProperties).offsetCalc(3),
       expected: {
-        fluid: 'calc((((100vw - var(--edge) * 2) / var(--columns) - (var(--gap) / var(--columns) * (var(--columns) - 1))) * 3) + var(--gap) * 3)',
-        full: 'calc((((90rem - var(--edge) * 2) / var(--columns) - (var(--gap) / var(--columns) * (var(--columns) - 1))) * 3) + var(--gap) * 3)',
+        // calc((((100vw - var(--edge) * 2) / var(--columns) - (var(--gap) / var(--columns) * (var(--columns) - 1))) * 3) + var(--gap) * 3)
+        fluid: 'calc((100vw - var(--edge)*2)/var(--columns)*3 - var(--gap)/var(--columns)*(var(--columns) - 1)*3 + var(--gap)*3)',
+        // calc((((90rem - var(--edge) * 2) / var(--columns) - (var(--gap) / var(--columns) * (var(--columns) - 1))) * 3) + var(--gap) * 3)
+        full: 'calc((90rem - var(--edge)*2)/var(--columns)*3 - var(--gap)/var(--columns)*(var(--columns) - 1)*3 + var(--gap)*3)',
       },
     },
   ],